+++ src/idc/class_lister_ignored.py
@@ -1,0 +1,9 @@
+from typing import List, Dict
+
+
+# classes that we don't want to be listed
+def list_classes() -> Dict[str, List[str]]:
+    return {
+        "seppl.io.Filter": [
+        ],
+    }