+++ CHANGES.rst
@@ -12,6 +12,7 @@ Changelog
 - the `metadata-from-name` filter can work on the path now as well (must be present)
 - the `label-present-is` filter now lets the user to specify a minimum number of pixels that need to be present in the layers
 - switched to `kasperl` library for base API and generic pipeline plugins
+- requiring seppl>=0.2.21 now
 
 
 0.0.13 (2025-07-15)